pyproject.toml
[project]
name = "repowire"
version = "0.1.0"
description = "Mesh network for Claude Code sessions - enables AI agents to communicate"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [{ name = "Prassanna Ravishankar" }]
keywords = ["ai", "claude", "mcp", "mesh", "agents", "tmux"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "mcp>=1.0.0",
    "python-socketio[asyncio_client]>=5.10.0",
    "libtmux>=0.37.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "pyyaml>=6.0.1",
    "rich>=13.7.0",
    "click>=8.1.7",
    "httpx>=0.26.0",
]

[project.optional-dependencies]
relay = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "redis>=5.0.1",
    "python-socketio>=5.10.0",
]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
]

[project.scripts]
repowire = "repowire.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["repowire"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.mypy]
python_version = "3.10"
strict = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
